+++ QuickGuidePanels/QuickGuidePanels.js
@@ -294,8 +294,9 @@ const QuickGuidePanelsBase = kind({
 					currentIcon={'circle'}
 					futureIcon={'circle'}
 					current={currentStep}
-					layout="quickGuidePanels"
+					layout="highlightCurrentStep"
 					total={totalSteps}
+					size={30}
 				/>
 			);
 		},